refactor: Replace complex AOP system with native Vapor middleware
This commit implements a critical architectural simplification that achieves
"elegant simplicity" by removing the custom Aspect-Oriented Programming (AOP)
system and replacing it with native Vapor middleware patterns.

Key Changes:
- Removed entire /Sources/App/Common/Aspects/ directory (5 files, 188+ lines)
- Created CorrelationIDMiddleware using native Vapor patterns
- Enhanced ErrorMiddleware with structured logging and error classification
- Updated Application-Setup.swift to use native middleware pipeline
- Removed AspectMiddleware and all AOP orchestration code
- Cleaned up aspect-related tests and imports

Architectural Benefits:
- Framework Harmony: Now works WITH Vapor conventions instead of against them
- Reduced Complexity: From 5-component AOP system to 1 simple middleware
- Better Performance: Native middleware execution path is more efficient
- Improved Maintainability: Standard Vapor patterns are easier to understand
- Preserved Functionality: All request tracing and error handling capabilities maintained

This refactoring follows the architectural vision principles of building less
but building it better, achieving power through simplicity rather than clever
complexity.

All tests pass - no functionality was lost in this simplification.
