+++ docs/architecture/ADRs/ADR-001-ServiceRegistry.md
