+++ docs/architecture/ADRs/ADR-002-Module-Colocation-and-Simplification.md
